[CDNG-2666]: harness secret manager and default org by event consumption (#19102)
* [CDNG-116]: project and org create delete event

* [CDNG-116]: no op producer

* [CDNG-116]: addressed comments

* [CDNG-116]: addressed comments

* [CDNG-116]: unit test fix

* [CDNG-116]: minor fixes

* [CDNG-116]: minor fixes

* [CDNG-116]: minor fixes

* [CDNG-116]: refactor

* [CDNG-116]: addressed comments

* [CDNG-116]: build fix

* [CDNG-2666]: initial changes

* [CDNG-2666]: initial changes

* [CDNG-2666]: test fix

* [CDNG-2666]: google kms name to uuid fix

* [CDNG-2666]: harness secret manager creation and deletion

* [CDNG-2666]: default org

* [CDNG-2666]: default org flag in aggregate dto

* [CDNG-2666]: major refactor

* [CDNG-2666]: undo extra changes

* [CDNG-2666]: undo extra changes

* [CDNG-2666]: bug fixes

* [CDNG-2666]: bug fixes

* [CDNG-2666]: build fix

* [CDNG-2666]: build fix

* [CDNG-2666]: bug fix

* [CDNG-2666]: build fix

* [CDNG-2666]: default org

* [CDNG-2666]: remove extra code

* [CDNG-2666]: addressed comments

+++ 120-ng-manager/src/main/java/io/harness/ng/core/remote/OrganizationResource.java
@@ -1,7 +1,9 @@
 package io.harness.ng.core.remote;
 
+import static io.harness.NGConstants.DEFAULT_ORG_IDENTIFIER;
 import static io.harness.annotations.dev.HarnessTeam.PL;
 import static io.harness.data.structure.EmptyPredicate.isEmpty;
+import static io.harness.exception.WingsException.USER;
 import static io.harness.ng.core.remote.OrganizationMapper.toResponseWrapper;
 import static io.harness.utils.PageUtils.getNGPageResponse;
 import static io.harness.utils.PageUtils.getPageRequest;
@@ -14,6 +16,7 @@ import io.harness.NGCommonEntityConstants;
 import io.harness.NGResourceFilterConstants;
 import io.harness.annotations.dev.OwnedBy;
 import io.harness.beans.SortOrder;
+import io.harness.exception.InvalidRequestException;
 import io.harness.ng.beans.PageRequest;
 import io.harness.ng.beans.PageResponse;
 import io.harness.ng.core.dto.ErrorDTO;
@@ -72,7 +75,10 @@ public class OrganizationResource {
   public ResponseDTO<OrganizationResponse> create(
       @NotNull @QueryParam(NGCommonEntityConstants.ACCOUNT_KEY) String accountIdentifier,
       @NotNull @Valid OrganizationRequest organizationDTO) {
-    organizationDTO.getOrganization().setHarnessManaged(organizationDTO.isHarnessManaged());
+    if (DEFAULT_ORG_IDENTIFIER.equals(organizationDTO.getOrganization().getIdentifier())) {
+      throw new InvalidRequestException(
+          String.format("%s cannot be used as org identifier", DEFAULT_ORG_IDENTIFIER), USER);
+    }
     Organization updatedOrganization = organizationService.create(accountIdentifier, organizationDTO.getOrganization());
     return ResponseDTO.newResponse(updatedOrganization.getVersion().toString(), toResponseWrapper(updatedOrganization));
   }
@@ -115,6 +121,12 @@ public class OrganizationResource {
       @NotNull @PathParam(NGCommonEntityConstants.IDENTIFIER_KEY) String identifier,
       @NotNull @QueryParam(NGCommonEntityConstants.ACCOUNT_KEY) String accountIdentifier,
       @NotNull @Valid OrganizationRequest organizationDTO) {
+    if (DEFAULT_ORG_IDENTIFIER.equals(identifier)) {
+      throw new InvalidRequestException(
+          String.format(
+              "Update operation not supported for Default Organization (identifier: [%s])", DEFAULT_ORG_IDENTIFIER),
+          USER);
+    }
     organizationDTO.getOrganization().setVersion(isNumeric(ifMatch) ? parseLong(ifMatch) : null);
     Organization updatedOrganization =
         organizationService.update(accountIdentifier, identifier, organizationDTO.getOrganization());
@@ -127,6 +139,12 @@ public class OrganizationResource {
   public ResponseDTO<Boolean> delete(@HeaderParam(IF_MATCH) String ifMatch,
       @NotNull @PathParam(NGCommonEntityConstants.IDENTIFIER_KEY) String identifier,
       @NotNull @QueryParam(NGCommonEntityConstants.ACCOUNT_KEY) String accountIdentifier) {
+    if (DEFAULT_ORG_IDENTIFIER.equals(identifier)) {
+      throw new InvalidRequestException(
+          String.format(
+              "Delete operation not supported for Default Organization (identifier: [%s])", DEFAULT_ORG_IDENTIFIER),
+          USER);
+    }
     return ResponseDTO.newResponse(
         organizationService.delete(accountIdentifier, identifier, isNumeric(ifMatch) ? parseLong(ifMatch) : null));
   }
